+++ components/camel-nats/src/test/java/org/apache/camel/component/nats/integration/NatsAuthTokenProducerIT.java
@@ -14,28 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.nats;
+package org.apache.camel.component.nats.integration;
 
-import java.io.IOException;
-
-import org.apache.camel.EndpointInject;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class NatsConsumerMaxMessagesQueueTest extends NatsTestSupport {
-
-    @EndpointInject("mock:result")
-    protected MockEndpoint mockResultEndpoint;
+public class NatsAuthTokenProducerIT extends NatsAuthITSupport {
 
     @Test
-    public void testMaxConsumer() throws InterruptedException, IOException {
-        mockResultEndpoint.setExpectedMessageCount(2);
-
-        template.sendBody("direct:send", "test");
-        template.sendBody("direct:send", "test1");
-
-        mockResultEndpoint.assertIsSatisfied();
+    public void sendTest() throws Exception {
+        template.sendBody("direct:send", "pippo");
     }
 
     @Override
@@ -44,10 +32,6 @@ public class NatsConsumerMaxMessagesQueueTest extends NatsTestSupport {
             @Override
             public void configure() throws Exception {
                 from("direct:send").to("nats:test");
-
-                from("nats:test?maxMessages=5&queueName=test").routeId("cons1").to(mockResultEndpoint);
-
-                from("nats:test?maxMessages=6&queueName=test").routeId("cons2").to(mockResultEndpoint);
             }
         };
     }
