+++ components/camel-nats/src/test/java/org/apache/camel/component/nats/integration/NatsConsumerMaxMessagesIT.java
@@ -14,24 +14,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.nats;
+package org.apache.camel.component.nats.integration;
+
+import java.io.IOException;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class NatsConsumerTest extends NatsTestSupport {
+public class NatsConsumerMaxMessagesIT extends NatsITSupport {
 
     @EndpointInject("mock:result")
     protected MockEndpoint mockResultEndpoint;
 
     @Test
-    public void testConsumer() throws Exception {
-        mockResultEndpoint.expectedBodiesReceived("Hello World");
-        mockResultEndpoint.expectedHeaderReceived(NatsConstants.NATS_SUBJECT, "test");
-
-        template.sendBody("direct:send", "Hello World");
+    public void testMaxConsumer() throws InterruptedException, IOException {
+        mockResultEndpoint.setExpectedMessageCount(5);
+        template.sendBody("direct:send", "test");
+        template.sendBody("direct:send", "test1");
+        template.sendBody("direct:send", "test2");
+        template.sendBody("direct:send", "test3");
+        template.sendBody("direct:send", "test4");
+        template.sendBody("direct:send", "test5");
+        template.sendBody("direct:send", "test6");
+        template.sendBody("direct:send", "test7");
+        template.sendBody("direct:send", "test8");
+        template.sendBody("direct:send", "test9");
+        template.sendBody("direct:send", "test10");
 
         mockResultEndpoint.assertIsSatisfied();
     }
@@ -41,9 +51,9 @@ public class NatsConsumerTest extends NatsTestSupport {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                from("direct:send").to("nats:test?flushConnection=true");
+                from("direct:send").to("nats:test");
 
-                from("nats:test?flushConnection=true").to(mockResultEndpoint);
+                from("nats:test?maxMessages=5").to(mockResultEndpoint);
             }
         };
     }
