+++ components/camel-nats/src/test/java/org/apache/camel/component/nats/integration/NatsProducerReplyToTimeoutIT.java
@@ -14,36 +14,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.nats;
+package org.apache.camel.component.nats.integration;
 
+import org.apache.camel.CamelExecutionException;
+import org.apache.camel.ExchangeTimedOutException;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
+import static org.apache.camel.test.junit5.TestSupport.assertIsInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 
-public class NatsProducerReplyToTest extends NatsTestSupport {
+public class NatsProducerReplyToTimeoutIT extends NatsITSupport {
 
     protected String startUri = "direct:start";
-    protected String middleUri = "nats:foo.middle";
+    protected String middleUri = "nats:foo.middle?requestTimeout=50";
     protected String resultUri = "mock:result";
 
-    protected MockEndpoint resultEndpoint;
     protected String body1 = "Camel";
-    protected String body2 = "World";
 
     @Test
-    public void testReplyTo() throws Exception {
-        resultEndpoint = context.getEndpoint(resultUri, MockEndpoint.class);
-        resultEndpoint.expectedBodiesReceivedInAnyOrder("Bye Camel", "Bye World");
-
-        String out = template.requestBody(startUri, body1, String.class);
-        String out2 = template.requestBody(startUri, body2, String.class);
-
-        resultEndpoint.assertIsSatisfied();
-
-        assertEquals("Bye Camel", out);
-        assertEquals("Bye World", out2);
+    public void testReplyToTimeout() {
+        try {
+            template.requestBody(startUri, body1, String.class);
+            fail("Should have thrown an exception");
+        } catch (CamelExecutionException e) {
+            ExchangeTimedOutException cause = assertIsInstanceOf(ExchangeTimedOutException.class, e.getCause());
+            assertEquals(50, cause.getTimeout());
+        }
     }
 
     @Override
@@ -53,6 +51,7 @@ public class NatsProducerReplyToTest extends NatsTestSupport {
                 from(startUri).to(middleUri).to(resultUri);
 
                 from(middleUri)
+                        .delayer(5000)
                         .transform(simple("Bye ${body}"));
             }
         };
